Fix: Calculate range correctly when class is exported (fixes #152)
The range of a class body should be from the opening brace to the
closing brace. To find the opening brace the ast converter parses over
modifiers of the class declaration and assumes it is the last token
after the last node modifier. This assumption is not always true when the
class is exported. This commit ensures we only skip over modifiers
which are after the class name or heritage clause.