Fix: Calculate range correctly for exported generic class (fixes #152)
When a class with a generic definition is exported we must take the
type parameters into account when finding the opening brace. This
continues the fix of pull request #153 and handles the case of exporting
a class with generics. We now find the last type parameter and use the
succeeding token as the opening brace.